+++ src/main/java/eu/europeana/metis/sandbox/entity/RecordEntity.java
@@ -1,0 +1,68 @@
+package eu.europeana.metis.sandbox.entity;
+
+import eu.europeana.metis.sandbox.common.Status;
+import eu.europeana.metis.sandbox.common.Step;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+
+@MappedSuperclass
+class RecordEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  protected Long id;
+
+  protected String recordId;
+
+  protected Integer datasetId;
+
+  @Enumerated(EnumType.STRING)
+  protected Step step;
+
+  @Enumerated(EnumType.STRING)
+  protected Status status;
+
+  public Long getId() {
+    return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+
+  public String getRecordId() {
+    return recordId;
+  }
+
+  public void setRecordId(String recordId) {
+    this.recordId = recordId;
+  }
+
+  public Integer getDatasetId() {
+    return datasetId;
+  }
+
+  public void setDatasetId(Integer datasetId) {
+    this.datasetId = datasetId;
+  }
+
+  public Step getStep() {
+    return step;
+  }
+
+  public void setStep(Step step) {
+    this.step = step;
+  }
+
+  public Status getStatus() {
+    return status;
+  }
+
+  public void setStatus(Status status) {
+    this.status = status;
+  }
+}
