+++ src/main/java/eu/europeana/metis/sandbox/entity/RecordLogEntity.java
@@ -3,30 +3,11 @@ package eu.europeana.metis.sandbox.entity;
 import eu.europeana.metis.sandbox.common.Status;
 import eu.europeana.metis.sandbox.common.Step;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.Table;
 
 @Entity
 @Table(name = "record_log")
-public class RecordLogEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  private String recordId;
-
-  private Integer datasetId;
-
-  @Enumerated(EnumType.STRING)
-  private Step step;
-
-  @Enumerated(EnumType.STRING)
-  private Status status;
+public class RecordLogEntity extends RecordEntity {
 
   private String content;
 
@@ -41,46 +22,6 @@ public class RecordLogEntity {
 
   public RecordLogEntity() {
     // provide explicit no-args constructor as it is required for Hibernate
-  }
-
-  public Long getId() {
-    return id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public String getRecordId() {
-    return recordId;
-  }
-
-  public void setRecordId(String recordId) {
-    this.recordId = recordId;
-  }
-
-  public Integer getDatasetId() {
-    return datasetId;
-  }
-
-  public void setDatasetId(Integer datasetId) {
-    this.datasetId = datasetId;
-  }
-
-  public Step getStep() {
-    return step;
-  }
-
-  public void setStep(Step step) {
-    this.step = step;
-  }
-
-  public Status getStatus() {
-    return status;
-  }
-
-  public void setStatus(Status status) {
-    this.status = status;
   }
 
   public String getContent() {
